Rename auth_controller -> user_controller
