+++ lib/views/auth_view.dart
@@ -6,7 +6,7 @@ import 'package:go_router/go_router.dart';
 import '../constants/privacy_notice.dart';
 import '../constants/terms_of_service.dart';
 import '../service_locator.dart';
-import '../controllers/auth_controller.dart';
+import '../controllers/user_controller.dart';
 import '../utils/config.dart';
 import '../widgets/auth/auth_form_fields.dart';
 import '../widgets/auth/auth_header.dart';
